requirements.txt
pytesseract
Pillow
python-dotenv